+++ src/parser/enrichers/trait/_module.mjs
@@ -16,3 +16,4 @@ export * as Orc from "./orc/_module.mjs";
 export * as Human from "./human/_module.mjs";
 export * as Tiefling from "./tiefling/_module.mjs";
 export * as Gnome from "./gnome/_module.mjs";
+export * as Eladrin from "./eladrin/_module.mjs";